voice_api/requirements.txt
amqp==2.5.2
asgiref==3.2.4
audioread==2.1.8
backcall==0.1.0
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.0.0
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
Django==3.0.4
djangorestframework==3.11.0
firebase==3.0.1
firebase-admin==4.0.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-speech==1.3.2
google-cloud-storage==1.26.0
google-cloud-texttospeech==0.5.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
kiwisolver==1.1.0
kombu==4.6.8
librosa==0.7.2
llvmlite==0.31.0
matplotlib==3.1.2
msgpack==1.0.0
noisereduce==1.1.0
numba==0.48.0
numpy==1.18.1
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydub==0.23.1
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
sqlparse==0.3.1
tqdm==4.42.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
zipp==3.1.0
